Make shure if user want to delete catalog

+++ src/routes/product/table/ProductTable.tsx
@@ -62,10 +62,18 @@ export const ProductTable = () => {
     const handleAction = (action: TActions, productId: number) => {
         switch (action) {
             case 'delete':
+                const curProduct = products.find(p => p.parent.id == productId);
+
+                if (curProduct?.parent) {
+                    if (!window.confirm(`Каталог ${curProduct.name} содержит дочерние продукты. вы уверены, что хотите удалить каталог ПОЛНОСТЬЮ?`)) {
+                        break;
+                    }
+                }
+
                 ProductService.delete(productId)
                     .then(() => {
                         setUpdatedProducts(updatedProducts.filter(product => product.id !== productId));
-                        alert('Продукт удален');
+                        alert('Продукт удален, перезагрузите страницу, что бы увидеть обновленную таблицу');
                     })
                 break;
